all: add in missing proto options and restructure package heirachy
Added in the missing proto options from
https://github.com/golang/protobuf/blob/882cf97a83ad205fd22af574246a3bc647d7a7d2/protoc-gen-go/descriptor/descriptor.proto#L318.

Restructure the layout so that each proto option (field, message, file,
etc) gets its own namespaces, and any language specific options are then
put in '<proto_type_option>/<lang>'.

Updates github.com/gunk/gunk/issues/104
