+++ file/cc/doc.go
@@ -1,0 +1,3 @@
+package cc
+
+// make this directory a Go package
