+++ file/csharp/doc.go
@@ -1,0 +1,3 @@
+package csharp
+
+// make this directory a Go package
