+++ file/doc.go
@@ -1,0 +1,3 @@
+package file
+
+// make this directory a Go package
