+++ file/java/doc.go
@@ -1,0 +1,3 @@
+package java
+
+// make this directory a Go package
