+++ file/objc/doc.go
@@ -1,0 +1,3 @@
+package objc
+
+// make this directory a Go package
