+++ file/php/doc.go
@@ -1,0 +1,3 @@
+package php
+
+// make this directory a Go package
