+++ file/ruby/doc.go
@@ -1,0 +1,3 @@
+package ruby
+
+// make this directory a Go package
