+++ file/swift/doc.go
@@ -1,3 +1,3 @@
-package opt
+package swift
 
 // make this directory a Go package
