--- opt.gunk
@@ -1,13 +1,0 @@
-// Package opt provides the standard gunk protobuf options.
-// 
-// Generated by the gunkconv tool.
-package opt
-
-// Deprecated is the deprecated option.
-type Deprecated bool
-
-// JavaPackage is the java_package option.
-type JavaPackage string
-
-// JavaMultipleFiles is the java_multiple_files option.
-type JavaMultipleFiles bool
